+++ sourcetracker/_cli/gibbs.py
@@ -114,6 +114,10 @@ from sourcetracker._util import parse_sample_metadata, biom_to_df
                     'sequences that contributed to a sink from a given '
                     'source. This option can be memory intensive if there are '
                     'a large number of features.'))
+@click.option('--sample_with_replacement', required=False,
+              default=False, show_default=True, is_flag=True,
+              help=('Sample with replacement instead of '
+                    'sample without replacement'))
 @click.option('--source_sink_column', required=False, default='SourceSink',
               type=click.STRING, show_default=True,
               help=('Sample metadata column indicating which samples should be'
@@ -133,8 +137,9 @@ from sourcetracker._util import parse_sample_metadata, biom_to_df
 def gibbs_cli(table_fp, mapping_fp, output_dir, loo, jobs, alpha1, alpha2,
               beta, source_rarefaction_depth, sink_rarefaction_depth, restarts,
               draws_per_restart, burnin, delay, cluster_start_delay,
-              per_sink_feature_assignments, source_sink_column,
-              source_column_value, sink_column_value, source_category_column):
+              per_sink_feature_assignments, sample_with_replacement,
+              source_sink_column, source_column_value,
+              sink_column_value, source_category_column):
     '''Gibb's sampler for Bayesian estimation of microbial sample sources.
 
     For details, see the project README file.
@@ -189,7 +194,8 @@ def gibbs_cli(table_fp, mapping_fp, output_dir, loo, jobs, alpha1, alpha2,
                              (source_rarefaction_depth, count_too_shallow,
                               shallowest))
         else:
-            csources = subsample_dataframe(csources, source_rarefaction_depth)
+            csources = subsample_dataframe(csources, source_rarefaction_depth,
+                                           replace=sample_with_replacement)
 
     # Prepare to rarify sink data if we are not doing LOO. If we are doing loo,
     # we skip the rarefaction, and set sinks to `None`.
@@ -207,7 +213,8 @@ def gibbs_cli(table_fp, mapping_fp, output_dir, loo, jobs, alpha1, alpha2,
                                  (sink_rarefaction_depth, count_too_shallow,
                                   shallowest))
             else:
-                sinks = subsample_dataframe(sinks, sink_rarefaction_depth)
+                sinks = subsample_dataframe(sinks, sink_rarefaction_depth,
+                                            replace=sample_with_replacement)
     else:
         sinks = None